librarian/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
websockets==12.0
pydantic==2.5.3
python-multipart==0.0.6
# qdrant-client==1.16.2  # Reserved for future AI chat feature
# sentence-transformers==2.7.0  # Reserved for future AI chat feature
# transformers==4.40.0  # Reserved for future AI chat feature
# torch==2.3.0  # Reserved for future AI chat feature - large NVIDIA dependency
numpy==1.26.4
asyncpg==0.29.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
redis==5.0.1
httpx==0.27.0
aiohttp==3.9.1
aiofiles==23.2.1
# llama-cpp-python==0.3.4  # Reserved for future AI chat feature
alembic==1.13.1
sqlalchemy==2.0.25
apprise==1.7.0
python-jose[cryptography]==3.3.0
aiosqlite==0.19.0
